+++ src/main/java/org/mule/module/hubspot/model/OAuthCredentials.java
@@ -11,13 +11,11 @@ package org.mule.module.hubspot.model;
 
 import java.io.Serializable;
 
-import org.mule.module.hubspot.client.HubSpotClient;
-
 public class OAuthCredentials implements Serializable {
 	
 	private static final long serialVersionUID = 1L;
 
-	private HubSpotClient client;
+	private String clientId;
 	private String userId;
 	private String accessToken;
 	private String expiresAt;
@@ -59,11 +57,11 @@ public class OAuthCredentials implements Serializable {
 		this.expiresAt = expiresAt;
 	}
 
-	public HubSpotClient getClient() {
-		return client;
+	public String getClientId() {
+		return clientId;
 	}
 
-	public void setClient(HubSpotClient client) {
-		this.client = client;
-	}		
+	public void setClientId(String clientId) {
+		this.clientId = clientId;
+	}
 }